+++ src/main/php/CallMap.php
@@ -43,9 +43,9 @@ class CallMap
             return false;
         }
 
-   #     if ($this->callMap[$method] instanceof InvocationResults) {
-   #         return $this->callMap[$method]->hasResultForInvocation($invocationCount - 1);
-   #     }
+        if ($this->callMap[$method] instanceof InvocationResults) {
+            return $this->callMap[$method]->hasResultForInvocation($invocationCount - 1);
+        }
 
         return true;
     }
@@ -65,7 +65,7 @@ class CallMap
             if (is_callable($this->callMap[$method])) {
                 return call_user_func_array($this->callMap[$method], $arguments);
             } elseif ($this->callMap[$method] instanceof InvocationResults) {
-                return $this->callMap[$method]->valueForInvocation($invocationCount - 1);
+                return $this->callMap[$method]->resultForInvocation($invocationCount - 1);
             } elseif ($this->callMap[$method] instanceof InvocationThrow) {
                 throw $this->callMap[$method]->exception();
             }